+++ Packages/com.unity.inputsystem/InputSystem/InputManager.cs
@@ -2556,6 +2556,20 @@ namespace UnityEngine.InputSystem
                 m_HaveSentStartupAnalytics = true;
             }
             #endif
+            
+            // See if we're supposed to only take events up to a certain time.
+            // NOTE: We do not require the events in the queue to be sorted. Instead, we will walk over
+            //       all events in the buffer each time. Note that if there are multiple events for the same
+            //       device, it depends on the producer of these events to queue them in correct order.
+            //       Otherwise, once an event with a newer timestamp has been processed, events coming later
+            //       in the buffer and having older timestamps will get rejected.
+
+            var currentTime = updateType == InputUpdateType.Fixed ? m_Runtime.currentTimeForFixedUpdate : m_Runtime.currentTime;
+            var timesliceEvents = shouldProcessInputEvents && InputSystem.settings.updateMode == InputSettings.UpdateMode.ProcessEventsInFixedUpdate;
+            
+            // mouse move compression will modify event buffer in-place
+            if (!settings.disableMouseMoveCompression)
+                CompressMouseMoveEvents.ProcessEvents(updateType, timesliceEvents ? currentTime : -1.0f, ref eventBuffer);
 
             ////TODO: manual mode must be treated like lockInputToGameView in editor
 
@@ -2573,16 +2587,6 @@ namespace UnityEngine.InputSystem
             InputStateBuffers.SwitchTo(m_StateBuffers, updateType);
 
             InputUpdate.OnUpdate(updateType);
-
-            // See if we're supposed to only take events up to a certain time.
-            // NOTE: We do not require the events in the queue to be sorted. Instead, we will walk over
-            //       all events in the buffer each time. Note that if there are multiple events for the same
-            //       device, it depends on the producer of these events to queue them in correct order.
-            //       Otherwise, once an event with a newer timestamp has been processed, events coming later
-            //       in the buffer and having older timestamps will get rejected.
-
-            var currentTime = updateType == InputUpdateType.Fixed ? m_Runtime.currentTimeForFixedUpdate : m_Runtime.currentTime;
-            var timesliceEvents = shouldProcessInputEvents && InputSystem.settings.updateMode == InputSettings.UpdateMode.ProcessEventsInFixedUpdate;
 
             // Early out if there's no events to process.
             if (eventBuffer.eventCount <= 0)