+++ Packages/com.unity.inputsystem/InputSystem/InputSettings.cs
@@ -470,6 +470,22 @@ namespace UnityEngine.InputSystem
                 OnChange();
             }
         }
+        
+        /// <summary>
+        /// Disables mouse move compression. Disable it if you want to get all mouse move events.
+        /// </summary>
+        public bool disableMouseMoveCompression
+        {
+            get => m_DisableMouseMoveCompression;
+            set
+            {
+                if (m_DisableMouseMoveCompression == value)
+                    return;
+
+                m_DisableMouseMoveCompression = value;
+                OnChange();
+            }
+        }
 
         /// <summary>
         /// List of device layouts used by the project.
@@ -541,6 +557,7 @@ namespace UnityEngine.InputSystem
         [SerializeField] private float m_DefaultHoldTime = 0.4f;
         [SerializeField] private float m_TapRadius = 5;
         [SerializeField] private float m_MultiTapDelayTime = 0.75f;
+        [SerializeField] private bool m_DisableMouseMoveCompression = false;
 
         internal void OnChange()
         {